Updates to the following Component: header, edit assignment and edit individual assignment
header: (1) Session wipe out, this is to prevent users from backing back upon logging out, thus users will be redirected back into the login page

Edit assignment: (1) assignments can now only be deleted when assignment ID is not tag in a assignment submissions entity (2) instructors are now able to filter via on going or new assignment with the current search filter.

Edit Individual Assignment: (1) resolve a bug when a user were to add a school instruments and level, would cause the drop-down select to not work accordingly as intended.