Updates to wage Component
Changes to the data presentation to allow better understanding of data for the activity log for the individual component.

added in filtering via school for wage

Changes from database filtering to single retrieval of data from the wage table database and filter the data locally. thus, this will expedite the process filter process for the end user for a better user experience